+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-07-25] [rework scoped shape](https://github.com/RubricLab/shapes/commit/76f70ee65df6547626cc37b03713b4e0d8b65d23)
 - [2025-07-24] [add gitignore](https://github.com/RubricLab/shapes/commit/898ca78f854d62764f83244e6abf8f8882cb9516)
 - [2025-07-18] [shapes core](https://github.com/RubricLab/shapes/commit/5180e86f867e3967c1db0597e288a5078e455a28)
 - [2025-07-18] [shapes core](https://github.com/RubricLab/shapes/commit/b68b864bcba279c5e0e2c706f19b8db96cfc51d9)